fixed huge security hole

+++ comments/delete.php
@@ -1,11 +1,13 @@
 <?php	
 	require_once '../database_access.php';
 
-echo "text";
+	session_start();
+	// make sure only the correct user can delete things
+	if(!isset($_SESSION['uid'])) die();
 
 	if (isset($_POST['comment_id'])) {
 		$comment = CommentQuery::create()->findPk($_POST['comment_id']);
-		if ($comment) {
+		if ($comment && $comment->getAuthorUserId() == $_SESSION['uid']) {
 			$comment->delete();
 			echo '<p>Successfully deleted comment</p>';
 		} else {
@@ -13,6 +15,5 @@ echo "text";
 		}
 			
 	}
-	
 
 ?>
